+++ resources/views/zapisi/neusag_proizod/admin/edit.blade.php
@@ -105,7 +105,7 @@
             </div>
 
             <br>
-            {!! Form::open(['method'=>'DELETE', 'action'=> ['NeusagController@destroy', $datas->id]]) !!}
+            {!! Form::open(['method'=>'DELETE', 'action'=> ['NeusagController@destroy', $datas->id], 'onsubmit'=>'return confirm("Da li ste sigurni?")']) !!}
             <div align="center">
                 {!! Form::submit('Obriši', ['class'=>'btn btn-danger']) !!}
             </div>
